Avoid deprecation warnings when using Gradle 6
In Gradle 5, configuring configProperties.config_loc has been
deprecated in favour of configDir which is passed through to
config_loc, while also providing better up-to-date checking.
Unfortunately, if the directory specified by configDir does not exist,
Gradle does not pass it through to config_loc. This means that
configDir can be set to gain better up-to-date-checking but that
config_loc must still be set explicitly. As a result, a deprecation
warnings is still generated when running with Gradle 5.

In Gradle 6, configDir has been replaced with configDirectory and the
value is always passed through to Checkstyle. This means that
configDirectory alone can be set, resulting in better up-to-date
checks and no deprecation warnings.